+++ R/mappers.R
@@ -1,3 +1,5 @@
+# Generics ----
+
 #' Constructors for `bru_mapper` objects
 #'
 #' @details
@@ -39,11 +41,7 @@ bru_mapper <- function(...) {
 #' @rdname bru_mapper_methods
 #' @name bru_mapper_methods
 ibm_n <- function(mapper, inla_f = FALSE, ...) {
-  if (!is.null(mapper[[".envir"]][["ibm_n"]])) {
-    mapper[[".envir"]][["ibm_n"]](mapper, inla_f = inla_f, ...)
-  } else {
-    UseMethod("ibm_n")
-  }
+  UseMethod("ibm_n")
 }
 
 #' @details
@@ -56,12 +54,8 @@ ibm_n <- function(mapper, inla_f = FALSE, ...) {
 #' that can accept `list()` inputs require their own methods implementations.
 #' @export
 #' @rdname bru_mapper_methods
-ibm_n_output <- function(mapper, input, inla_f = FALSE, ...) {
-  if (!is.null(mapper[[".envir"]][["ibm_n"]])) {
-    mapper[[".envir"]][["ibm_n"]](mapper, input, inla_f = inla_f, ...)
-  } else {
-    UseMethod("ibm_n_output")
-  }
+ibm_n_output <- function(mapper, input, state = NULL, inla_f = FALSE, ...) {
+  UseMethod("ibm_n_output")
 }
 
 
@@ -74,11 +68,7 @@ ibm_n_output <- function(mapper, input, inla_f = FALSE, ...) {
 #' @export
 #' @rdname bru_mapper_methods
 ibm_values <- function(mapper, inla_f = FALSE, ...) {
-  if (!is.null(mapper[[".envir"]][["ibm_values"]])) {
-    mapper[[".envir"]][["ibm_values"]](mapper, inla_f = inla_f, ...)
-  } else {
-    UseMethod("ibm_values")
-  }
+  UseMethod("ibm_values")
 }
 #' @details
 #' * `ibm_amatrix` Generic, will become deprecated in 2.7.0. Use `ibm_jacobian`
@@ -89,12 +79,12 @@ ibm_values <- function(mapper, inla_f = FALSE, ...) {
 #' @export
 #' @rdname bru_mapper_methods
 ibm_amatrix <- function(mapper, input, state = NULL, inla_f = FALSE, ...) {
-  if (!is.null(mapper[[".envir"]][["ibm_amatrix"]])) {
-    mapper[[".envir"]][["ibm_amatrix"]](
-      mapper, input, state = state, inla_f = inla_f, ...)
-  } else {
-    UseMethod("ibm_amatrix")
-  }
+  lifecycle::deprecate_soft(
+    "2.6.0.9000",
+    "ibm_amatrix()",
+    "ibm_jacobian()"
+  )
+  UseMethod("ibm_amatrix")
 }
 
 #' @details
@@ -106,12 +96,7 @@ ibm_amatrix <- function(mapper, input, state = NULL, inla_f = FALSE, ...) {
 #' @export
 #' @rdname bru_mapper_methods
 ibm_is_linear <- function(mapper, ...) {
-  if (!is.null(mapper[[".envir"]][["ibm_is_linear"]])) {
-    mapper[[".envir"]][["ibm_is_linear"]](
-      mapper = mapper, ...)
-  } else {
-    UseMethod("ibm_is_linear")
-  }
+  UseMethod("ibm_is_linear")
 }
 
 #' @details
@@ -124,12 +109,7 @@ ibm_is_linear <- function(mapper, ...) {
 #' @export
 #' @rdname bru_mapper_methods
 ibm_jacobian <- function(mapper, input, state = NULL, inla_f = FALSE, ...) {
-  if (!is.null(mapper[[".envir"]][["ibm_jacobian"]])) {
-    mapper[[".envir"]][["ibm_jacobian"]](
-      mapper = mapper, input = input, state = state, inla_f = inla_f, ...)
-  } else {
-    UseMethod("ibm_jacobian")
-  }
+  UseMethod("ibm_jacobian")
 }
 
 
@@ -146,12 +126,7 @@ ibm_jacobian <- function(mapper, input, state = NULL, inla_f = FALSE, ...) {
 #' @export
 #' @rdname bru_mapper_methods
 ibm_linear <- function(mapper, input, state = NULL, ...) {
-  if (!is.null(mapper[[".envir"]][["ibm_linear"]])) {
-    mapper[[".envir"]][["ibm_linear"]](
-      mapper, input, state = state, ...)
-  } else {
-    UseMethod("ibm_linear")
-  }
+  UseMethod("ibm_linear")
 }
 
 #' @details
@@ -165,13 +140,28 @@ ibm_linear <- function(mapper, input, state = NULL, ...) {
 #' @export
 #' @rdname bru_mapper_methods
 ibm_eval <- function(mapper, input, state = NULL, ...) {
-  if (!is.null(mapper[[".envir"]][["ibm_eval"]])) {
-    mapper[[".envir"]][["ibm_eval"]](
-      mapper, input, state = state, ...)
-  } else {
-    UseMethod("ibm_eval")
-  }
+  UseMethod("ibm_eval")
 }
+
+#' @details
+#' * `ibm_names` Generic.
+#' Implementations must return a character vector of sub-mapper names, or `NULL`.
+#' Intended for providing information about multi-mappers and mapper collections.
+#' @export
+#' @rdname bru_mapper_methods
+ibm_names <- function(mapper) {
+  UseMethod("ibm_names")
+}
+
+#' @param value a character vector of the same length as the number
+#' of sub-mappers in the mapper
+#' * `ibm_names<-` Generic.
+#' @export
+#' @rdname bru_mapper_methods
+`ibm_names<-` <- function(mapper, value) {
+  UseMethod("ibm_names<-")
+}
+
 
 #' @details
 #' * `ibm_inla_subset` Generic.
@@ -184,12 +174,7 @@ ibm_eval <- function(mapper, input, state = NULL, ...) {
 #' @export
 #' @rdname bru_mapper_methods
 ibm_inla_subset <- function(mapper, ...) {
-  if (!is.null(mapper[[".envir"]][["ibm_inla_subset"]])) {
-    mapper[[".envir"]][["ibm_inla_subset"]](
-      mapper, ...)
-  } else {
-    UseMethod("ibm_inla_subset")
-  }
+  UseMethod("ibm_inla_subset")
 }
 #' @details
 #' * `ibm_valid_input` Generic.
@@ -199,74 +184,210 @@ ibm_inla_subset <- function(mapper, ...) {
 #' @export
 #' @rdname bru_mapper_methods
 ibm_valid_input <- function(mapper, input, inla_f = FALSE, ...) {
-  if (!is.null(mapper[[".envir"]][["ibm_valid_input"]])) {
-    mapper[[".envir"]][["ibm_valid_input"]](
-      mapper, input, inla_f = inla_f, ...)
-  } else {
-    UseMethod("ibm_valid_input")
+  UseMethod("ibm_valid_input")
+}
+
+
+
+# Summaries ----
+
+ibm_shortname <- function(mapper, ...) {
+  shortname <- sub("^bru_mapper_", "", class(mapper)[1])
+  if (nchar(shortname) == 0) {
+    shortname <- "default"
   }
+  shortname
+}
+
+
+#' @title mapper object summaries
+#'
+#' @param object `bru_mapper` object to summarise
+#' @param \dots Unused arguments
+#' @param prefix character prefix for each line. Default `""`.
+#' @param initial character prefix for the first line. Default `initial=prefix`.
+#' @param depth The recursion depth for multi/collection/pipe mappers.
+#' Default 1, to only show the collection, and not the contents of the sub-mappers.
+#' @export
+#' @method summary bru_mapper
+#' @rdname bru_mapper_summary
+summary.bru_mapper <- function(object, ...,
+                               prefix = "",
+                               initial = prefix,
+                               depth = 1) {
+  summary_object <- paste0(initial, ibm_shortname(object))
+  class(summary_object) <- c("summary_bru_mapper",
+                             class(summary_object))
+  summary_object
+}
+
+#' @export
+#' @method summary bru_mapper_multi
+#' @rdname bru_mapper_summary
+summary.bru_mapper_multi <- function(object, ...,
+                                     prefix = "",
+                                     initial = prefix,
+                                     depth = 1) {
+  summary_object <- NextMethod()
+  if (depth <= 0) {
+    class(summary_object) <- c("summary_bru_mapper", class(summary_object))
+    return(summary_object)
+  }
+  sub_prefix <- paste0(prefix, "      ")
+  summary_object <-
+    paste0(
+      summary_object,
+      "(",
+      paste0(
+        vapply(names(object[["mappers"]]),
+               function(nm) paste0(nm,
+                                   " = ",
+                                   summary(
+                                     object[["mappers"]][[nm]],
+                                     prefix = sub_prefix,
+                                     initial = "",
+                                     depth = depth - 1)
+                                   ),
+               ""),
+        collapse = ", "),
+      ")"
+    )
+  class(summary_object) <- c("summary_bru_mapper", class(summary_object))
+  summary_object
+}
+
+#' @export
+#' @method summary bru_mapper_pipe
+#' @rdname bru_mapper_summary
+summary.bru_mapper_pipe <- function(object, ...,
+                                    prefix = "",
+                                    initial = prefix,
+                                    depth = 1) {
+  summary_object <- NextMethod()
+  if (depth <= 0) {
+    class(summary_object) <- c("summary_bru_mapper", class(summary_object))
+    return(summary_object)
+  }
+  sub_prefix <- paste0(prefix, "      ")
+  summary_object <-
+    paste0(
+      summary_object,
+      " = ",
+      paste0(
+        vapply(object[["mappers"]],
+               function(x) {
+                 summary(x,
+                         prefix = sub_prefix,
+                         initial = "",
+                         depth = depth - 1)
+               },
+               ""),
+        collapse = paste0(" -> ")
+      )
+    )
+  class(summary_object) <- c("summary_bru_mapper", class(summary_object))
+  summary_object
+}
+
+#' @export
+#' @method summary bru_mapper_collect
+#' @rdname bru_mapper_summary
+summary.bru_mapper_collect <- function(object, ...,
+                                       prefix = "",
+                                       initial = prefix,
+                                       depth = 1) {
+  summary_object <- NextMethod()
+  if (depth <= 0) {
+    class(summary_object) <- c("summary_bru_mapper", class(summary_object))
+    return(summary_object)
+  }
+  sub_prefix <- paste0(prefix, "      ")
+  summary_object <-
+    paste0(
+      summary_object,
+      "(",
+      paste0(
+        vapply(seq_along(object[["mappers"]]),
+               function(k) {
+                 nm <- names(object[["mappers"]])[k]
+                 if (is.null(nm)) {
+                   nm <- as.character(k)
+                 }
+                 paste0(nm,
+                        " = ",
+                        summary(
+                          object[["mappers"]][[k]],
+                          prefix = sub_prefix,
+                          initial = "",
+                          depth = depth - 1
+                        ),
+                        if (object[["hidden"]] && (k > 1)) {
+                          "(hidden)"
+                        } else {
+                          NULL
+                        }
+                 )
+               },
+               ""),
+        collapse = ", "),
+      ")"
+    )
+  class(summary_object) <- c("summary_bru_mapper", class(summary_object))
+  summary_object
+}
+
+
+#' @param x Object to be printed
+#' @export
+#' @method print summary_bru_mapper
+#' @rdname bru_mapper_summary
+print.summary_bru_mapper <- function(x, ...) {
+  cat(x)
+  invisible(x)
 }
 
 
 # MAPPERS ----
+## Constructor ----
 
 #' @details * `bru_mapper_define` adds the `new_class` and "bru_mapper" class
 #' names to the inheritance list for the input `mapper` object, unless the object
 #' already inherits from these.
-#' If provided, mapper method functions are added to an environment
-#' `.envir` in the object.  The generic methods look for these
-#' functions first,
-#' and otherwise call `UseMethod()`.  This is an alternative to using `.S3method()`
+#'
+#' To register mapper classes and methods in scripts, use `.S3method()`
 #' to register the methods, e.g.
 #' `.S3method("ibm_jacobian", "my_mapper_class", ibm_jacobian.my_mapper_class)`.
 #' @param mapper For `bru_mapper_define`, a prototype mapper object, see Details.
 #' For `bru_mapper_scale`, a mapper to be scaled.
 #' @param new_class If non-`NULL`, this is added at the front of the class definition
-#' @param methods Optional `list` of named method definitions; See Details.
 #' @param \dots Deprecated, alternative way to supply optional method definitions.
+#' @param methods Deprecated.
 #'
 #' @export
 #' @rdname bru_mapper
 bru_mapper_define <- function(mapper,
                               new_class = NULL,
-                              methods = NULL,
-                              ...) {
-  valid_method_names <-
-    c(
-      "ibm_n",
-      "ibm_n_output",
-      "ibm_values",
-      "ibm_amatrix",
-      "ibm_is_linear",
-      "ibm_jacobian",
-      "ibm_linear",
-      "ibm_eval",
-      "ibm_inla_subset",
-      "ibm_valid_input"
+                              ...,
+                              methods = NULL) {
+  if (!is.null(methods)) {
+    txt <-
+      c("In packages with Suggests: inlabru, add method information, e.g.:",
+        "  #' @rawNamespace if (getRversion() >= '3.6.0') {",
+        "  #'   S3method(inlabru::bru_get_mapper, inla_rspde)",
+        "  #'   S3method(inlabru::ibm_n, bru_mapper_inla_rspde)",
+        "  #'   S3method(inlabru::ibm_values, bru_mapper_inla_rspde)",
+        "  #'   S3method(inlabru::ibm_jacobian, bru_mapper_inla_rspde)",
+        "  #' }")
+    lifecycle::deprecate_warn(
+      when = "2.6.0.9000",
+      what = "bru_mapper_define(methods)",
+      details =
+        c(
+          "In scripts, use '.S3method()' to register each method.",
+          paste(txt, sep = "\n"),
+          "In packages with Imports: inlabru, `#' @export` should be sufficient."
+        )
     )
-  if (any(c(
-    valid_method_names,
-    "ibm_n_inla",
-    "ibm_values_inla",
-    "ibm_amatrix_inla",
-    "ibm_valid_input_inla"
-  ) %in%
-    names(list(...)))) {
-    warning(
-      paste0(
-        "Deprecated use of named method arguments for 'bru_mapper'.\n",
-        "Use methods = list(methodname = ..., ...) instead."
-      )
-    )
-    method_names <-
-      setdiff(
-        intersect(
-          valid_method_names,
-          names(list(...))
-        ),
-        names(methods)
-      )
-    methods <- c(list(...)[method_names], methods)
   }
   if (!inherits(mapper, "bru_mapper")) {
     class(mapper) <- c("bru_mapper", class(mapper))
@@ -277,28 +398,10 @@ bru_mapper_define <- function(mapper,
   if (is.null(mapper[[".envir"]])) {
     mapper$.envir <- new.env()
   }
-  if (!is.null(methods)) {
-    deprecated_methods <- setdiff(names(methods),
-                                  valid_method_names)
-    if (length(deprecated_methods) > 0) {
-      warning(
-        paste0(
-          "Unknown bru_mapper method names detected, ignoring methods\n",
-          paste0(deprecated_methods, collapse = ", "),
-          "."
-        )
-      )
-    }
-    for (method in setdiff(names(methods), deprecated_methods)) {
-      if (!is.null(methods[[method]])) {
-        assign(method,
-               methods[[method]],
-               envir = mapper[[".envir"]])
-      }
-    }
-  }
   mapper
 }
+
+## Default methods ----
 
 #' @details * `bru_mapper.default` calls `bru_mapper_define`, passing all
 #' arguments along. Mapper implementations should call [bru_mapper_define()]
@@ -307,8 +410,10 @@ bru_mapper_define <- function(mapper,
 #' @export
 #' @rdname bru_mapper
 bru_mapper.default <- function(...) {
-  # TODO: Mark deprecated from version 2.7.0
-  # .Deprecated("bru_mapper_define")
+  lifecycle::deprecate_warn(
+    "2.6.0.9000",
+    "bru_mapper.default()",
+    "bru_mapper_define()")
   bru_mapper_define(...)
 }
 
@@ -358,7 +463,7 @@ ibm_n.default <- function(mapper, inla_f = FALSE, ...) {
 
 #' @export
 #' @rdname bru_mapper_methods
-ibm_n_output.default <- function(mapper, input, inla_f = FALSE, ...) {
+ibm_n_output.default <- function(mapper, input, state = NULL, inla_f = FALSE, ...) {
   NROW(input)
 }
 
@@ -460,14 +565,21 @@ ibm_eval.default <- function(mapper, input, state = NULL, ...) {
     stop("Non-linear mappers must implement their own ibm_eval() method.")
   }
 
-  val <- numeric(ibm_n_output(mapper, input, ...))
+  val <- numeric(ibm_n_output(mapper, input, state = state, ...))
 
   if ((ibm_n(mapper) > 0) && !is.null(state)) {
     A <- ibm_jacobian(mapper, input = input, state = state, ...)
-    val + A %*% state
+    val <- val + A %*% state
   }
 
   val
+}
+
+
+#' @export
+#' @rdname bru_mapper_methods
+ibm_names.default <- function(mapper, ...) {
+  NULL
 }
 
 
@@ -771,7 +883,7 @@ ibm_n.bru_mapper_taylor <- function(mapper, inla_f = FALSE, multi = FALSE, ...) 
 
 #' @export
 #' @rdname bru_mapper_methods
-ibm_n_output.bru_mapper_taylor <- function(mapper, input, inla_f = FALSE, ...) {
+ibm_n_output.bru_mapper_taylor <- function(mapper, input, ...) {
   mapper[["n_output"]]
 }
 
@@ -1149,12 +1261,16 @@ bru_mapper_scale <- function(mapper = NULL, ...) {
   }
 }
 
+#' @param n_state integer giving the length of the state vector for mappers
+#' that have state dependent output size.
 #' @export
 #' @rdname bru_mapper_methods
-ibm_n.bru_mapper_scale <- function(mapper, ..., n_state = NULL) {
+ibm_n.bru_mapper_scale <- function(mapper, ..., state = NULL, n_state = NULL) {
   if (is.null(mapper[["mapper"]])) {
     # Output size depends on the state size
-    if (is.null(n_state)) {
+    if (!is.null(state)) {
+      length(state)
+    } else if (is.null(n_state)) {
       NA_integer_
     } else {
       n_state
@@ -1165,23 +1281,27 @@ ibm_n.bru_mapper_scale <- function(mapper, ..., n_state = NULL) {
 }
 #' @export
 #' @rdname bru_mapper_methods
-ibm_n_output.bru_mapper_scale <- function(mapper, input, ..., n_state = NULL) {
+ibm_n_output.bru_mapper_scale <- function(mapper, input, state = NULL, ...,
+                                          n_state = NULL) {
   if (is.null(mapper[["mapper"]])) {
-    if (is.null(n_state)) {
+    if (!is.null(state)) {
+      length(state)
+    } else if (is.null(n_state)) {
       # To allow scalar input weights, do not assume NROW(input) size
       NA_integer_
     } else {
       n_state
     }
   } else {
-    ibm_n_output(mapper[["mapper"]], input, ...)
+    ibm_n_output(mapper[["mapper"]], input, state = state, ..., n_state = n_state)
   }
 }
 #' @export
 #' @rdname bru_mapper_methods
-ibm_values.bru_mapper_scale <- function(mapper, ..., n_state = NULL) {
+ibm_values.bru_mapper_scale <- function(mapper, ...,
+                                        state = NULL, n_state = NULL) {
   if (is.null(mapper[["mapper"]])) {
-    n_state <- ibm_n(mapper, n_state = n_state)
+    n_state <- ibm_n(mapper, state = state, n_state = n_state)
     if (is.na(n_state)) {
       # Don't know how big the mapper will be
       NULL
@@ -1238,7 +1358,15 @@ ibm_jacobian.bru_mapper_scale <- function(mapper, input, state = NULL, ...,
 #' @rdname bru_mapper_methods
 ibm_linear.bru_mapper_scale <- function(mapper, input, state, ...) {
   if (is.null(mapper[["mapper"]])) {
-    stop("bru_mapper_scale can only be linearised as part of a bru_mapper_pipe")
+    stopifnot(!is.null(state))
+    return(
+      bru_mapper_taylor(
+        offset = ibm_eval(mapper, input, state, ...),
+        jacobian = ibm_jacobian(mapper, input, state, ...),
+        state0 = state,
+        values_mapper = mapper
+      )
+    )
   }
 
   # Behaviour for v2.6.0, non-null mapper
@@ -1260,6 +1388,9 @@ ibm_eval.bru_mapper_scale <- function(mapper, input, state = NULL, ...,
                                       sub_lin = NULL) {
   if (is.null(mapper[["mapper"]])) {
     stopifnot(!is.null(state))
+    if (is.null(input)) {
+      return(state)
+    }
     scale <- as.vector(input)
     ok <- !is.na(scale)
     scale[!ok] <- 0
@@ -1312,6 +1443,12 @@ bru_mapper_pipe <- function(mappers, ...) {
   n_multi <- vapply(mappers, function(x) as.integer(ibm_n(x)), 0L)
   n <- ibm_n(mappers[[1]])
   is_linear <- all(is_linear_multi)
+  if (is.null(names(mappers))) {
+    names(mappers) <- as.character(seq_along(mappers))
+  } else if ("" %in% names(mappers)) {
+    warning("Either all or none of the pipe sub-mappers should be named.", immediate. = TRUE)
+    names(mappers) <- as.character(seq_along(mappers))
+  }
   bru_mapper_define(list(mappers = mappers,
                          is_linear_multi,
                          is_linear = is_linear,
@@ -1323,7 +1460,7 @@ bru_mapper_pipe <- function(mappers, ...) {
 #' @export
 #' @rdname bru_mapper_methods
 ibm_n.bru_mapper_pipe <- function(mapper, ..., state = NULL) {
-  if (is.na(mapper[["mappers"]][[1]]) && is.null(state)) {
+  if (is.null(mapper[["mappers"]][[1]]) && is.null(state)) {
     return(NA_integer_)
   }
   ibm_n(mapper[["mappers"]][[1]], ..., state = state)
@@ -1346,7 +1483,13 @@ ibm_values.bru_mapper_pipe <- function(mapper, ...) {
 #' @rdname bru_mapper_methods
 ibm_jacobian.bru_mapper_pipe <- function(mapper, input, state = NULL, ...) {
   state_k <- state
-  for (k in seq_along(mapper[["mappers"]])) {
+  if (!is.null(input)) {
+    if (is.null(names(input))) {
+      names(input) <- names(mapper[["mappers"]])[seq_along(input)]
+    }
+  }
+  first <- names(mapper[["mappers"]])[1]
+  for (k in names(mapper[["mappers"]])) {
     # TODO: Introduce an "eval and Jacobian" method to avoid
     # double-computing the evaluations when needing both.
     A_k <- ibm_jacobian(mapper[["mappers"]][[k]],
@@ -1356,7 +1499,7 @@ ibm_jacobian.bru_mapper_pipe <- function(mapper, input, state = NULL, ...) {
                         input = input[[k]],
                         state = state_k,
                         ...)
-    if (k == 1) {
+    if (k == first) {
       A <- A_k
     } else {
       A <- A_k %*% A
@@ -1402,7 +1545,7 @@ ibm_eval.bru_mapper_pipe <- function(mapper, input, state = NULL, ...) {
 #' @rdname bru_mapper_methods
 ibm_valid_input.bru_mapper_pipe <- function(mapper, input, ..., multi = FALSE) {
   valid <- lapply(
-    seq_len(mapper[["mappers"]]),
+    seq_along(mapper[["mappers"]]),
     function(k) {
       ibm_valid_input(mapper[["mappers"]][[k]], input[[k]], ...)
     })
@@ -1534,7 +1677,7 @@ bru_mapper_multi_prepare_input <- function(mapper, input) {
 #' @details
 #' * `ibm_jacobian` for `bru_mapper_multi` accepts a list with
 #' named entries, or a list with unnamed but ordered elements.
-#' The names must match the sub-mappers, see [names.bru_mapper_multi()].
+#' The names must match the sub-mappers, see [ibm_names.bru_mapper_multi()].
 #' Each list element should take a format accepted by the corresponding
 #' sub-mapper. In case each element is a vector, the input can be given as a
 #' data.frame with named columns, a matrix with named columns, or a matrix
@@ -1656,7 +1799,7 @@ bm_multi_indexing <- function(mapper, input) {
 #' @details
 #' * `ibm_valid_input` for `bru_mapper_multi` accepts a list with
 #' named entries, or a list with unnamed but ordered elements.
-#' The names must match the sub-mappers, see [names.bru_mapper_multi()].
+#' The names must match the sub-mappers, see [ibm_names.bru_mapper_multi()].
 #' Each list element should take a format accepted by the corresponding
 #' sub-mapper. In case each element is a vector, the input can be given as a
 #' data.frame with named columns, a matrix with named columns, or a matrix
@@ -1738,21 +1881,21 @@ ibm_valid_input.bru_mapper_multi <- function(mapper, input,
 #' sub-mappers list
 #' @export
 #' @rdname bru_mapper_methods
-`names.bru_mapper_multi` <- function(x) {
-  names(x[["mappers"]])
+`ibm_names.bru_mapper_multi` <- function(mapper) {
+  names(mapper[["mappers"]])
 }
 
 #' @param value a character vector of up to the same length as the number
 #' of mappers in the multi-mapper x
 #' @export
 #' @rdname bru_mapper_methods
-`names<-.bru_mapper_multi` <- function(x, value) {
-  names(x[["mappers"]]) <- value
-  names(x[["n_multi"]]) <- value
-  names(x[["n_inla_multi"]]) <- value
-  names(x[["values_multi"]]) <- value
-  names(x[["values_inla_multi"]]) <- value
-  x
+`ibm_names<-.bru_mapper_multi` <- function(mapper, value) {
+  names(mapper[["mappers"]]) <- value
+  names(mapper[["n_multi"]]) <- value
+  names(mapper[["n_inla_multi"]]) <- value
+  names(mapper[["values_multi"]]) <- value
+  names(mapper[["values_inla_multi"]]) <- value
+  mapper
 }
 
 
@@ -1822,6 +1965,7 @@ bm_collect_indexing <- function(mapper, input) {
 #' @export
 #' @rdname bru_mapper_methods
 ibm_n_output.bru_mapper_collect <- function(mapper, input,
+                                            state = NULL,
                                             inla_f = FALSE,
                                             multi = FALSE, ...) {
   if (mapper[["hidden"]] && inla_f) {
@@ -1930,7 +2074,7 @@ bm_collect_sub_lin <- function(mapper, input, state,
 #' @details
 #' * `ibm_jacobian` for `bru_mapper_collect` accepts a list with
 #' named entries, or a list with unnamed but ordered elements.
-#' The names must match the sub-mappers, see [names.bru_mapper_collect()].
+#' The names must match the sub-mappers, see [ibm_names.bru_mapper_collect()].
 #' Each list element should take a format accepted by the corresponding
 #' sub-mapper. In case each element is a vector, the input can be given as a
 #' data.frame with named columns, a matrix with named columns, or a matrix
@@ -1977,9 +2121,9 @@ ibm_eval.bru_mapper_collect <- function(mapper, input, state,
 
   if (!multi) {
     # Combine the vectors (b1, b2, b3, ...) -> c(b1, b2, b3, ...)
-    off <- do.call(c, off)
+    val <- do.call(c, val)
   }
-  off
+  val
 }
 
 
@@ -2014,7 +2158,7 @@ ibm_linear.bru_mapper_collect <- function(mapper, input, state,
 #' @details
 #' * `ibm_valid_input` for `bru_mapper_collect` accepts a list with
 #' named entries, or a list with unnamed but ordered elements.
-#' The names must match the sub-mappers, see [names.bru_mapper_collect()].
+#' The names must match the sub-mappers, see [ibm_names.bru_mapper_collect()].
 #' Each list element should take a format accepted by the corresponding
 #' sub-mapper. In case each element is a vector, the input can be given as a
 #' data.frame with named columns, a matrix with named columns, or a matrix
@@ -2098,19 +2242,19 @@ ibm_valid_input.bru_mapper_collect <- function(mapper, input, inla_f = FALSE, mu
 #' sub-mappers list
 #' @export
 #' @rdname bru_mapper_methods
-`names.bru_mapper_collect` <- function(x) {
-  names(x[["mappers"]])
+`ibm_names.bru_mapper_collect` <- function(mapper) {
+  names(mapper[["mappers"]])
 }
 
-#' @param value a character vector of up to the same length as x
 #' @export
 #' @rdname bru_mapper_methods
-`names<-.bru_mapper_collect` <- function(x, value) {
-  names(x[["mappers"]]) <- value
-  names(x[["n_multi"]]) <- value
-  names(x[["values_multi"]]) <- value
-  x
+`ibm_names<-.bru_mapper_collect` <- function(mapper, value) {
+  names(mapper[["mappers"]]) <- value
+  names(mapper[["n_multi"]]) <- value
+  names(mapper[["values_multi"]]) <- value
+  mapper
 }
+