+++ tests/testthat/test_mapper.R
@@ -1,5 +1,44 @@
 local_bru_testthat_setup()
 
+test_that("Pipe mapper", {
+  skip_on_cran()
+  local_bru_safe_inla()
+  mapper <-
+    bru_mapper_pipe(
+      list(
+        bru_mapper_multi(list(
+          space = bru_mapper_index(4),
+          time = bru_mapper_index(3)
+        )),
+        bru_mapper_scale()
+      )
+    )
+
+  expect_equal(ibm_n(mapper), 12)
+  expect_equal(ibm_n(mapper, multi = TRUE),
+               list(space = 4, time = 3))
+  expect_equal(ibm_values(mapper), seq_len(12))
+  expect_equal(
+    as.data.frame(ibm_values(mapper, multi = TRUE)),
+    expand.grid(space = seq_len(4), time = seq_len(3)),
+    ignore_attr = TRUE
+  )
+
+  olist_data <-
+    list(
+      list(space = 2:4, time = 1:3),
+      2
+    )
+  A <- Matrix::sparseMatrix(
+    i = 1:3,
+    j = c(2, 7, 12),
+    x = 1 * 2,
+    dims = c(3, 12)
+  )
+  expect_equal(ibm_jacobian(mapper, olist_data), A)
+})
+
+
 test_that("Multi-mapper bru input", {
   skip_on_cran()
   local_bru_safe_inla()
@@ -35,23 +74,23 @@ test_that("Multi-mapper bru input", {
 
   data <- cbind(df_data, y = 1:3)
 
-  cmp1 <- y ~ indep(list(time = time, space = space),
-    model = "ar1", mapper = mapper
+  cmp1 <- y ~ indep(list(space = space, time = time),
+                    model = "ar1", mapper = mapper
   ) - 1
   fit1 <- bru(cmp1, family = "gaussian", data = data)
 
   cmp2 <- y ~ indep(list(space, time),
-    model = "ar1", mapper = mapper
+                    model = "ar1", mapper = mapper
   ) - 1
   fit2 <- bru(cmp2, family = "gaussian", data = data)
 
   cmp3 <- y ~ indep(data.frame(time = time, space = space),
-    model = "ar1", mapper = mapper
+                    model = "ar1", mapper = mapper
   ) - 1
   fit3 <- bru(cmp3, family = "gaussian", data = data)
 
   cmp4 <- y ~ indep(cbind(space, time),
-    model = "ar1", mapper = mapper
+                    model = "ar1", mapper = mapper
   ) - 1
   fit4 <- bru(cmp4, family = "gaussian", data = data)
 
@@ -61,64 +100,16 @@ test_that("Multi-mapper bru input", {
 })
 
 
-# Define in outer environment:
 
 
 test_that("User defined mappers", {
-  # User defined mapper objects
-
-  ibm_jacobian.bm_test <- function(mapper, input, ...) {
-    message("---- IBM_JACOBIAN from inner environment ----")
-    Matrix::sparseMatrix(
-      i = seq_along(input),
-      j = input,
-      x = rep(2, length(input)),
-      dims = c(length(input), mapper$n)
-    )
-  }
-
-  bm_test <- function(n, ...) {
-    bru_mapper_define(
-      list(n = n),
-      new_class = "bm_test",
-      methods = list(
-        ibm_jacobian = ibm_jacobian.bm_test
-      )
-    )
-  }
-
-  m <- bm_test(n = 20)
-  cmp <- y ~ -1 + indep(x, model = "iid", mapper = m)
-  mydata <- data.frame(y = rnorm(15) + 2 * (1:15), x = 1:15)
-  expect_message(
-    object = {
-      ibm_jacobian(m, mydata$x)
-    },
-    "---- IBM_JACOBIAN from inner environment ----",
-    label = "ibm_jacobian generic call"
-  )
-
-  skip_on_cran()
-  local_bru_safe_inla()
-
-  expect_message(
-    object = {
-      fit <- bru(cmp, data = mydata, family = "gaussian")
-    },
-    "---- IBM_JACOBIAN from inner environment ----",
-    label = "Non-interactive bru() call"
-  )
-})
-
-
-test_that("User defined mappers 2", {
   # .S3method was unavailable in R 3.6!
   skip_if_not(utils::compareVersion("4", R.Version()$major) <= 0)
 
   # User defined mapper objects
 
   ibm_jacobian.bm_test <- function(mapper, input, ...) {
-    message("---- IBM_JACOBIAN from inner environment 2 ----")
+    message("---- IBM_JACOBIAN from inner environment ----")
     Matrix::sparseMatrix(
       i = seq_along(input),
       j = input,
@@ -142,7 +133,7 @@ test_that("User defined mappers 2", {
     object = {
       ibm_jacobian(m, mydata$x)
     },
-    "---- IBM_JACOBIAN from inner environment 2 ----",
+    "---- IBM_JACOBIAN from inner environment ----",
     label = "ibm_jacobian generic call"
   )
 
@@ -153,7 +144,7 @@ test_that("User defined mappers 2", {
     object = {
       fit <- bru(cmp, data = mydata, family = "gaussian")
     },
-    "---- IBM_JACOBIAN from inner environment 2 ----",
+    "---- IBM_JACOBIAN from inner environment ----",
     label = "Non-interactive bru() call"
   )
 })